fixed redux
