+++ verification/curator-service/ui/src/redux/filters/selectors.ts
@@ -1,0 +1,3 @@
+import { RootState } from '../store';
+
+export const countryList = (state: RootState) => state.countryListReducer.countryList;
