+++ verification/curator-service/ui/src/redux/filters/slice.ts
@@ -1,0 +1,18 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+const initialState = {
+    countryList: [],
+};
+
+const slice = createSlice({
+    name: 'countryList',
+    initialState,
+    reducers: {
+        setCountries(state: any, action) {
+            state.countryList = action.payload;
+        },
+    },
+});
+
+export const { setCountries } = slice.actions;
+export default slice.reducer;
