+++ verification/curator-service/ui/src/redux/store.ts
@@ -2,7 +2,7 @@ import { configureStore, combineReducers } from '@reduxjs/toolkit';
 
 import appReducer from './app/slice';
 import authReducer from './auth/slice';
-import countryListReducer from './countryList/CountryList';
+import countryListReducer from './filters/slice';
 
 export const rootReducer = combineReducers({
     app: appReducer,
